requirements.txt
apache-airflow-providers-amazon==9.8.0
apache-airflow-providers-apache-spark==5.3.0
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-papermill==3.11.2
beautifulsoup4==4.13.4
confluent-kafka==2.12.1
duckdb==1.2.2
Flask==2.2.5
# kafka-python==2.2.15
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
praw==7.8.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==2.3.7
xlrd==2.0.1